Added another version of url shortener that works with a PostgreSQL database.
