+++ url_shortener_v2/paths.csv
@@ -1,0 +1,5 @@
+path,url
+/urlshort-godoc,https://godoc.org/github.com/gophercises/urlshort
+/yaml-godoc,https://godoc.org/gopkg.in/yaml.v2
+/urlshort,https://github.com/gophercises/urlshort
+/urlshort-final,https://github.com/gophercises/urlshort/tree/solution
